+++ src/main/java/com/goorm/wordsketch/entity/VocabularyType.java
@@ -1,4 +1,4 @@
-package com.goorm.wordsketch.vocabulary;
+package com.goorm.wordsketch.entity;
 
 public enum VocabularyType {
