+++ src/main/java/com/goorm/wordsketch/repository/VocabularyRepository.java
@@ -2,7 +2,8 @@ package com.goorm.wordsketch.repository;
 
 import com.goorm.wordsketch.entity.Vocabulary;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-public interface VocabularyRepository extends JpaRepository<Vocabulary, Integer> {
-
+@Repository
+public interface VocabularyRepository extends JpaRepository<Vocabulary, Long> {
 }
