--- src/main/java/com/goorm/wordsketch/vocabulary/VocabularyRepository.java
@@ -1,8 +1,0 @@
-package com.goorm.wordsketch.vocabulary;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface VocabularyRepository extends JpaRepository<Vocabulary, Long> {
-}
